Change progress bar type from int to double. ab#813 #1266

+++ src/Client/Shared/Components/FileBrowser/FileBrowser.razor.cs
@@ -36,7 +36,7 @@ public partial class FileBrowser
     // ProgressBar
     private string ProgressBarCurrentText { get; set; } = default!;
     private string ProgressBarCurrentSubText { get; set; } = default!;
-    private int ProgressBarCurrentValue { get; set; }
+    private double ProgressBarCurrentValue { get; set; }
     private int ProgressBarMax { get; set; }
     private CancellationTokenSource? _progressBarCts;
